feat: update makefile

+++ makefile
@@ -1,8 +1,8 @@
 start-docker:
-	docker compose --env-file .env -f ./docker/docker-compose.yml up -d
+	clear && docker compose --env-file .env -f ./docker/docker-compose.yml up -d
 
 stop-docker:
-	docker compose --env-file .env -f ./docker/docker-compose.yml down --remove-orphans
+	clear && docker compose --env-file .env -f ./docker/docker-compose.yml down --remove-orphans
 
 clean-docker:
 	clear && \
